Get rid of strsep in cache.c, avoid useless copy
This destructive parsing is bad, as shown by the fact that its mere
use in the code makes tuning the parser a hell.  So let's drop it
before going further.

This also removes the need to memcpy the buffer, since we won't be
modifying it.

Note: there are also loads of strsep() calls in cvs_direct.c to get rid of.